updated navbar to either user/guest, need export
