+++ yardsalelocator/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\index.js":"1","C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\App.js":"2","C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\reportWebVitals.js":"3","C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\components\\Navbar.js":"4"},{"size":500,"mtime":499162500000,"results":"5","hashOfConfig":"6"},{"size":285,"mtime":1611080621727,"results":"7","hashOfConfig":"6"},{"size":362,"mtime":499162500000,"results":"8","hashOfConfig":"6"},{"size":1141,"mtime":1611079707386,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"f3pq9n",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\index.js",[],["19","20"],"C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\App.js",[],"C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\reportWebVitals.js",[],"C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\components\\Navbar.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\index.js":"1","C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\App.js":"2","C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\reportWebVitals.js":"3","C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\components\\Navbar.jsx":"4"},{"size":500,"mtime":499162500000,"results":"5","hashOfConfig":"6"},{"size":287,"mtime":1611083343236,"results":"7","hashOfConfig":"6"},{"size":362,"mtime":499162500000,"results":"8","hashOfConfig":"6"},{"size":2598,"mtime":1611083425683,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"f3pq9n",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\index.js",[],["19","20"],"C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\App.js",[],"C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\reportWebVitals.js",[],"C:\\Users\\razer\\Desktop\\Bethel\\Final Project\\team-java-boys-frontend\\yardsalelocator\\src\\components\\Navbar.jsx",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
