+++ yardsalelocator/src/App.js
@@ -2,6 +2,8 @@ import Navbar from "./components/Navbar";
 import {BrowserRouter as Router} from 'react-router-dom';
 import './App.css';
 
+
+
 function App() {
   return (
     <div>
